# AI-Assignments/realTime-foodDelivery-service/API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: SwiftEats API
  version: 1.0.0
  description: |
    API specification for **SwiftEats** — a real-time food delivery platform.

    ## Architecture

    Six Spring Boot microservices share domain code in `platform-lib`. **All external clients**
    (UIs, Postman, curl) call **backend-service** at `http://localhost:8080`. The gateway
    proxies domain routes to internal services.

    | Service | Port | Routes |
    |---------|------|--------|
    | backend-service | 8080 | `/auth/*`, public `/restaurants/*`, `/drivers/*`, order tracking |
    | entities-service | 8081 | `/admin/restaurants/*` |
    | order-service | 8082 | `/orders/*`, `/admin/orders/*` |
    | payment-service | 8083 | `/payments/*` |
    | refund-service | 8084 | `/refunds/*` |
    | analytics-service | 8085 | `/analytics/*`, `/admin/analytics/*` |

    **Unified Swagger UI:** http://localhost:8080/swagger-ui.html (dropdown for each service spec).

    ## Modules

    - **Auth** — customer register, login, profile (backend-service)
    - **Restaurant** — browse, search, menu cache (backend); admin CRUD (entities-service)
    - **Order** — idempotent creation, async payment, state machine (order-service)
    - **Payment** — RabbitMQ worker, mock gateway (payment-service)
    - **Refund** — initiation → successful mock refund (refund-service)
    - **Tracking** — Kafka GPS ingest, Redis hot path, SSE (backend-service)
...

    **Order state flow:**
...

...

...

...

...

...

  contact:
...
  license:
...
servers:
...
paths:
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new customer account
      operationId: registerCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterCustomerRequest'
      responses:
        '201':
          description: Account created; returns profile and API token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerAuthResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with email or phone + password
      operationId: loginCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginCustomerRequest'
      responses:
        '200':
          description: Authenticated; returns profile and rotated API token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerAuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Get current customer profile
      operationId: getCustomerProfile
      security:
        - CustomerIdAuth: []
        - CustomerApiKeyAuth: []
      responses:
        '200':
          description: Customer profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/profile:
    patch:
      tags: [Auth]
      summary: Update current customer profile
      operationId: updateCustomerProfile
      security:
        - CustomerIdAuth: []
        - CustomerApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerProfileRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/v1/orders:
    post:
      tags: [Orders]
      summary: Create order (async payment)
      description: |
        Accepts order and returns **202** immediately. Payment is processed asynchronously
        via outbox + RabbitMQ. Use the same `Idempotency-Key` to safely retry.
      operationId: createOrder
      security:
        - CustomerIdAuth: []
          CustomerApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '202':
          description: Order accepted; payment processing asynchronously
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderAcceptedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    get:
      tags: [Orders]
      summary: List current customer's orders
      operationId: listCustomerOrders
      security:
        - CustomerIdAuth: []
        - CustomerApiKeyAuth: []
      parameters:
        - name: scope
          in: query
          schema:
            type: string
            enum: [active, history, all]
            default: all
          description: |
            `active` — in-progress orders; `history` — delivered/cancelled/failed; `all` — both.
      responses:
        '200':
          description: Orders newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order details
      operationId: getOrder
      security:
        - CustomerIdAuth: []
          CustomerApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order with line items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/orders/{orderId}/history:
    get:
      tags: [Orders]
      summary: Get order state transition history
      operationId: getOrderHistory
      security:
        - CustomerIdAuth: []
          CustomerApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: State history entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderStateHistoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/analytics/delays:
    get:
      tags: [Analytics]
      summary: Delay root-cause breakdown for a city and date (UC1)
      operationId: analyzeDelaysByCity
      parameters:
        - name: city
          in: query
          required: true
          schema: { type: string, example: Pune }
        - name: date
          in: query
          required: true
          schema: { type: string, format: date, example: '2025-03-17' }
      responses:
        '200':
          description: Delay analysis with narrative and recommendations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DelayAnalysisResponse'
        '412':
          description: Sample data not imported
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/analytics/failures:
    get:
      tags: [Analytics]
      summary: Client failure analysis for a date range (UC2)
      operationId: analyzeFailuresByClient
      parameters:
        - name: clientId
          in: query
          required: true
          schema: { type: integer, format: int64, example: 337 }
        - name: from
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Failure breakdown with correlated fleet and warehouse signals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailureAnalysisResponse'
  /api/v1/analytics/failures/by-warehouse:
    get:
      tags: [Analytics]
      summary: Warehouse-linked failure analysis (UC3)
      operationId: analyzeFailuresByWarehouse
      parameters:
        - name: warehouseId
          in: query
          required: true
          schema: { type: integer, format: int64, example: 2 }
        - name: monthParam
          in: query
          description: Year-month as yyyy-MM (alternative to year + month)
          schema: { type: string, example: '2025-08' }
        - name: year
          in: query
          schema: { type: integer }
        - name: month
          in: query
          schema: { type: integer, minimum: 1, maximum: 12 }
      responses:
        '200':
          description: Warehouse failure analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseFailureResponse'
  /api/v1/analytics/failures/compare:
    get:
      tags: [Analytics]
      summary: Compare failure causes between two cities (UC4)
      operationId: compareCityFailures
      parameters:
        - name: cityA
          in: query
          required: true
          schema: { type: string, example: Pune }
        - name: cityB
          in: query
          required: true
          schema: { type: string, example: Mumbai }
        - name: monthParam
          in: query
          schema: { type: string, example: '2025-08' }
        - name: year
          in: query
          schema: { type: integer }
        - name: month
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: City comparison
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityComparisonResponse'
  /api/v1/analytics/capacity-projection:
    get:
      tags: [Analytics]
      summary: Project warehouse capacity risk for onboarding (UC6)
      operationId: projectCapacityRisk
      parameters:
        - name: clientId
          in: query
          required: true
          schema: { type: integer, format: int64 }
        - name: additionalMonthlyOrders
          in: query
          required: true
          schema: { type: integer, example: 20000 }
      responses:
        '200':
          description: Capacity risk report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapacityRiskResponse'
  /api/v1/analytics/insights/query:
    post:
      tags: [Analytics]
      summary: Structured insight query across all use cases (UC1–UC6)
      operationId: executeInsightQuery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsightQueryRequest'
      responses:
        '200':
          description: Narrative insight with evidence and recommendations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsightResponse'
...
components:
  securitySchemes:
    AdminApiKey:
      type: apiKey
      in: header
      name: X-Admin-Api-Key
    CustomerIdAuth:
      type: apiKey
      in: header
      name: X-Customer-Id
      description: Customer UUID; required together with X-Customer-Api-Key on protected routes
    CustomerApiKeyAuth:
      type: apiKey
      in: header
      name: X-Customer-Api-Key
      description: Per-customer API token from the customer record; required together with X-Customer-Id
    DriverAuth:
      type: apiKey
      in: header
      name: X-Driver-Api-Key
      description: Must match the driver UUID in the request path
  parameters:
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string }
      description: Client-generated unique key for safe retries
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not authorized for this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Resource already exists (e.g. duplicate email)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, timestamp, path]
      properties:
        code: { type: string, example: NOT_FOUND }
        message: { type: string }
        timestamp: { type: string, format: date-time }
        path: { type: string }
    RegisterCustomerRequest:
      type: object
      required: [name, email, phone, password, addressLine1, city]
      properties:
        name:
          type: string
          description: Stored with `[T] ` prefix (e.g. `[T] Jane Doe`)
          example: Jane Doe
        addressLine1: { type: string }
        addressLine2: { type: string }
        city: { type: string }
        state: { type: string, example: Maharashtra }
        pincode: { type: string }
    LoginCustomerRequest:
      type: object
      required: [loginId, password]
      properties:
        loginId:
          type: string
          description: Email address or 10-digit phone number
        password: { type: string, minLength: 8 }
    UpdateCustomerProfileRequest:
      type: object
      required: [name, email, phone, addressLine1, city]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string, pattern: '^[0-9]{10}$' }
        addressLine1: { type: string }
        addressLine2: { type: string }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
    CustomerProfileResponse:
      type: object
      properties:
        customerId: { type: string, format: uuid }
        name:
          type: string
          example: "[T] Demo Customer"
        email: { type: string, format: email }
        phone: { type: string }
        addressLine1: { type: string }
        addressLine2: { type: string }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CustomerAuthResponse:
      type: object
      properties:
        profile:
          $ref: '#/components/schemas/CustomerProfileResponse'
        customerId: { type: string, format: uuid }
        apiToken: { type: string }
...
    CreateOrderRequest:
      type: object
      required: [restaurantId, deliveryAddress, items]
      properties:
        restaurantId: { type: string, format: uuid }
        deliveryAddress: { type: string }
        city: { type: string }
        state: { type: string }
        pincode: { type: string }
        paymentMode: { type: string, example: UPI }
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [menuItemId, quantity]
            properties:
              menuItemId: { type: string, format: uuid }
              quantity: { type: integer, minimum: 1 }
    OrderSummaryResponse:
      type: object
      properties:
        orderId: { type: string, format: uuid }
        restaurantId: { type: string, format: uuid }
        restaurantName: { type: string }
        status: { type: string }
        paymentStatus: { type: string }
        totalAmount: { type: number, format: double }
        deliveryAddressLine1: { type: string }
        city: { type: string }
        createdAt: { type: string, format: date-time }
    OrderAcceptedResponse:
      type: object
      properties:
        orderId: { type: string, format: uuid }
        status:
          type: string
          enum: [PENDING_PAYMENT]
        paymentStatus:
          type: string
          enum: [PENDING]
        totalAmount: { type: number, format: double }
        message: { type: string }
    OrderItemResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        menuItemId: { type: string, format: uuid }
        name:
          type: string
          example: "[T] Kolhapuri Misal"
        unitPrice: { type: number }
        lineTotal: { type: number }
    OrderResponse:
      type: object
      properties:
        orderId: { type: string, format: uuid }
        customerId: { type: string, format: uuid }
        restaurantId: { type: string, format: uuid }
        driverId: { type: string, format: uuid, nullable: true }
        status:
          type: string
          enum:
            - PENDING_PAYMENT
            - CONFIRMED
            - PREPARING
            - OUT_FOR_DELIVERY
            - DELIVERED
            - PAYMENT_FAILED
            - DELAYED
            - FAILED
            - CANCELLED
            - RETURNED
        paymentStatus:
          type: string
          enum: [PENDING, SUCCESS, FAILED]
        totalAmount: { type: number }
        deliveryAddressLine1: { type: string }
        city: { type: string }
        failureReason: { type: string, nullable: true }
        delayReason: { type: string, nullable: true }
        promisedDeliveryAt: { type: string, format: date-time, nullable: true }
        prepStartedAt: { type: string, format: date-time, nullable: true }
        outForDeliveryAt: { type: string, format: date-time, nullable: true }
        deliveredAt: { type: string, format: date-time, nullable: true }
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        createdAt: { type: string, format: date-time }
    OrderStateHistoryResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        fromStatus: { type: string, nullable: true }
        toStatus: { type: string }
        changedBy: { type: string }
        reason: { type: string, nullable: true }
        changedAt: { type: string, format: date-time }
...
    InsightQueryRequest:
      type: object
      required: [queryType, parameters]
      properties:
        queryType:
          type: string
          enum:
            - DELAY_BY_CITY
            - FAILURES_BY_CLIENT
            - FAILURES_BY_WAREHOUSE
            - COMPARE_CITIES
            - FESTIVAL_ANALYSIS
            - CAPACITY_PROJECTION
        parameters:
          type: object
          additionalProperties: { type: string }
    InsightResponse:
      type: object
      properties:
        narrative: { type: string }
        recommendations:
          type: array
          items: { type: string }
        evidence:
          type: object
          additionalProperties: true
    DelayAnalysisResponse:
      type: object
      properties:
        city: { type: string }
        date: { type: string, format: date }
        totalAffected: { type: integer }
        failureReasonCounts:
          type: object
          additionalProperties: { type: integer, format: int64 }
        heavyTrafficCount: { type: integer, format: int64 }
        slowPackingCount: { type: integer, format: int64 }
        negativeFeedbackCount: { type: integer, format: int64 }
        narrative: { type: string }
        recommendations:
          type: array
          items: { type: string }
    FailureAnalysisResponse:
      type: object
      properties:
        clientId: { type: integer, format: int64 }
        from: { type: string }
        to: { type: string }
        totalFailed: { type: integer }
        breakdown:
          type: array
          items:
            type: object
            properties:
              failureReason: { type: string }
              count: { type: integer, format: int64 }
              fleetIssues: { type: string, nullable: true }
              warehouseIssues: { type: string, nullable: true }
        narrative: { type: string }
        recommendations:
          type: array
          items: { type: string }
    WarehouseFailureResponse:
      type: object
      properties:
        warehouseId: { type: integer, format: int64 }
        warehouseName: { type: string }
        month: { type: integer }
        year: { type: integer }
        totalFailed: { type: integer }
        failureReasonCounts:
          type: object
          additionalProperties: { type: integer, format: int64 }
        warehouseNoteCounts:
          type: object
          additionalProperties: { type: integer, format: int64 }
        narrative: { type: string }
        recommendations:
          type: array
          items: { type: string }
    CityComparisonResponse:
      type: object
      properties:
        cityA: { type: string }
        cityB: { type: string }
        month: { type: string }
        cityAFailures:
          type: object
          additionalProperties: { type: integer, format: int64 }
        cityBFailures:
          type: object
          additionalProperties: { type: integer, format: int64 }
        narrative: { type: string }
        recommendations:
          type: array
          items: { type: string }
    CapacityRiskResponse:
      type: object
      properties:
        clientId: { type: integer, format: int64 }
        clientName: { type: string }
        additionalMonthlyOrders: { type: integer }
        historicalFailureRate: { type: number, format: double }
        warehouseRisks:
          type: array
          items:
            type: object
            properties:
              warehouseId: { type: integer, format: int64 }
              warehouseName: { type: string }
              city: { type: string }
              capacity: { type: integer }
              currentMonthlyOrders: { type: integer, format: int64 }
              projectedMonthlyOrders: { type: integer, format: int64 }
              projectedUtilizationPct: { type: number, format: double }
              highRisk: { type: boolean }
        narrative: { type: string }
        recommendations:
          type: array
          items: { type: string }